Wpf refactor restructure
