+++ src/binstarjs03.AerialOBJ.WpfApp/UIElements/Components/INumericBox.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace binstarjs03.AerialOBJ.WpfApp.UIElements.Components;
+
 public interface INumericBox
 {
     public void ClearValue();
