+++ src/binstarjs03.AerialOBJ.WpfApp/UIElements/Windows/DebugLogWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Windows;
+﻿using System.ComponentModel;
+using System.Windows;
 
 namespace binstarjs03.AerialOBJ.WpfApp.UIElements.Windows;
 
@@ -8,5 +9,11 @@ public partial class DebugLogWindow : Window
     {
         InitializeComponent();
         DataContext = new DebugLogWindowVM(this);
+    }
+
+    protected override void OnClosing(CancelEventArgs e)
+    {
+        e.Cancel = true;
+        Hide();
     }
 }
