+++ src/binstarjs03.AerialOBJ.WpfApp/UIElements/Windows/MainWindow.xaml.cs
@@ -33,6 +33,7 @@ public partial class MainWindow : Window
 
     protected override void OnClosing(CancelEventArgs e)
     {
+        // TODO do proper cleanup before exiting
         Environment.Exit(0);
         //Application.Current.Shutdown();
     }
